backend/requirements.txt
django==4.2.3
django-taggit==4.0.0